Refactor to make CpuMetricsCollector, CpuStats, and metrics submission OS-agnostic, create separate crates for Windows and Linux for reading raw CPU data
