+++ crates/datadog-metrics-collector/src/lib.rs
@@ -10,3 +10,7 @@
 pub mod azure_instance;
 pub mod azure_tags;
 pub mod cpu;
+#[cfg(not(target_os = "windows"))]
+pub mod linux;
+#[cfg(target_os = "windows")]
+pub mod windows;
